♻️ refactor:  UrlLauncherまわりのリファクタリング
