+++ lib/application/url_launcher/state/url_launch_data.dart
@@ -3,25 +3,16 @@
 // found in the LICENSE file.
 
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 part 'url_launch_data.freezed.dart';
-
-/// URL起動モード
-enum UrlLaunchMode {
-  /// Leaves the decision of how to launch the URL to the platform
-  /// implementation.
-  platformDefault,
-
-  /// Loads the URL in an in-app web view (e.g., Safari View Controller).
-  inAppWebView,
-}
 
 /// URL起動データ
 @freezed
 class UrlLaunchData with _$UrlLaunchData {
   const factory UrlLaunchData({
     required String urlString,
-    required UrlLaunchMode mode,
+    required LaunchMode mode,
   }) = _UrlLaunchData;
 
   const UrlLaunchData._();
